ml-services/biobert-service/pyproject.toml
[tool.poetry]
name = "biolens-biobert-service"
version = "0.1.0"
description = "BioBERT service for medical entity extraction and symptom analysis"
authors = ["BioLens Team"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
transformers = "^4.36.0"
torch = "^2.1.0"
numpy = "^1.24.0"
scikit-learn = "^1.3.0"
pydantic = "^2.5.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
black = "^23.11.0"
isort = "^5.12.0"
hypothesis = "^6.92.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 88